+++ pdfforms/__init__.py
@@ -1,0 +1,4 @@
+"inspect and fill pdf fillable forms"
+from .pdfforms import fill_pdfs, inspect_pdfs
+from .transforms import *
+from .version import __version__